do day 3 puzzles

+++ Advent2020/Day01_Sums3.cs
@@ -2,6 +2,6 @@
 {
     public class Day01_Sums3 : Day01_Sums2
     {
-        public override int GetResult() => GetProduct(SampleValues, 2020, 3);
+        public override long GetResult() => GetProduct(SampleValues, 2020, 3);
     }
 }
